maps: Defer getting preallocate flags until open
The initial implementation here attempted to set the preallocate map
flags in the map initialization functions for each map, however these
were not guaranteed to be run after the commandline argument is parsed
and set into pkg/option.

Defer the preallocate flag option fetch until OpenOrCreate(), which
should be called by OpenAfterMount(), which is run in the daemon's
initEnv() function after the new preallocation setting is configured.

Signed-off-by: Joe Stringer <joe@cilium.io>